// src/pages/studio/types/types.ts
// Data Grid page
//---------------

export type DataGridRowType = {
    id: number;
    avatar: string;
    full_name: string;
    post: string;
    email: string;
    title: string;
    post_update: string;
    salary: number;
    age: string;
    experience: string;
    status: number;
    feedTypes : number[];
}


export type DataVideosGridRowType = {
    id: number;
    avatar: string;
    full_name: string;
    post: string;
    videoUrl: string;
    title: string;
    last_update: string;
    salary: number;
    age: string;
    experience: string;
    status: number;
}